Assignment 1
#+STARTUP: showall
#+TITLE: Program 000:
#+SUBTITLE: Operating Systems
#+AUTHOR: Brian C. Ladd
#+DATE: Fall 2017

* Getting Started
This assignment is for you to practice putting a file on your QPi, compile it, and run it.

** Download =mystery.s=
For this assignment, the source code is provided in the assignment directory. The file, =mystery.s= is in the same folder as this assignment file. You should download it.

** Install =qemu= and set up a QPi
=qemu= is an emulator program that can emulate an ARM processor on a x86 platform (or any other platform on which it can run). You should install it as outlined in [[../../notes/qpi.org]] The instructions are for a Linux machine. You can find information on running a Raspberry Pi emulator on Window on line with a search for those terms.

** Copy the mystery program onto your account on the QPi
Once your QPi boots, copy =mystery.s= into a folder in your home directory. Note that the QPi is running a version of Raspbian, a Linux derive